requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.6.12

tqdm
pillow

intel_extension_for_pytorch